+++ src/components/FeatureFlagsUI.tsx
@@ -1,43 +1,122 @@
 /* eslint-disable @typescript-eslint/no-empty-function */
-/* eslint-disable no-console */
 import React, { ReactElement } from 'react';
-import CoreUI, { FeatureFlagsUIProps } from './CoreUI';
-
 import {
   useEditFeatureFlag,
   useGetFeatures,
   useResetFeatureFlags,
 } from './index';
+import NoFlags from './NoFlags';
+import ReadOnly from './ReadOnly';
 
-const FeatureFlagsUI = ({
+type CoreUIType = {
+  onFeatureChange?: (id: string, checked: boolean) => void;
+  onFeatureReset?: () => void;
+  persist?: boolean;
+  readonly?: boolean;
+  notDefaultIndicator?: ReactElement;
+};
+
+// type FeatureFlagsUIProps = {
+//   onFeatureChange?: (flagId?: string, isActive?: boolean) => void;
+//   onFeatureReset?: () => void;
+//   readonly?: boolean;
+//   notDefaultIndicator?: ReactElement;
+// };
+
+const CoreUI = ({
   onFeatureChange = () => {},
   onFeatureReset = () => {},
+  persist = false,
   readonly = false,
-  notDefaultIndicator,
-}: FeatureFlagsUIProps): ReactElement => {
+  notDefaultIndicator = (
+    <span
+      className='badge badge-pill badge-info rounded-pill bg-info text-dark'
+      data-testid='notDefaultIndicatorDefault'
+    >
+      Changed
+    </span>
+  ),
+}: CoreUIType): ReactElement => {
   const features = useGetFeatures();
   const editFeature = useEditFeatureFlag();
   const resetFlags = useResetFeatureFlags();
-  const persist = false; // KKD - change me
+  if (!features || features.length === 0) {
+    return <NoFlags />;
+  }
 
-  return (
+  return readonly ? (
+    <ReadOnly
+      features={features || []}
+      notDefaultIndicator={notDefaultIndicator}
+    />
+  ) : (
     <>
-      <CoreUI
-        persist={persist}
-        features={features}
-        onFeatureClick={(id, checked) => {
-          editFeature(id, checked);
-          onFeatureChange(id, checked);
-        }}
-        onFeatureReset={() => {
-          resetFlags();
-          onFeatureReset();
-        }}
-        readonly={readonly}
-        notDefaultIndicator={notDefaultIndicator}
-      />
+      <ul data-testid='coreFeatureFlagsUI'>
+        {features.map((feature) => (
+          <li key={feature.id}>
+            <div className='form-check form-switch custom-control custom-switch'>
+              <input
+                className='form-check-input custom-control-input'
+                type='checkbox'
+                id={feature.id}
+                checked={feature.active}
+                onChange={(e) => {
+                  editFeature(feature.id, e.target.checked);
+                  onFeatureChange(feature.id, e.target.checked);
+                }}
+                disabled={readonly}
+              />{' '}
+              <label
+                className='form-check-label custom-control-label font-weight-bold fw-bold'
+                htmlFor={feature.id}
+              >
+                {feature.title ?? feature.id}
+                {feature.active !== feature.original ? (
+                  <span data-testid='flagNotInitialWarning'>
+                    {' '}
+                    {notDefaultIndicator}
+                  </span>
+                ) : null}
+              </label>
+              {feature.description ? (
+                <div
+                  className='text-muted font-italic fst-italic'
+                  data-testid='feature_description'
+                >
+                  {feature.description}
+                </div>
+              ) : null}
+            </div>
+          </li>
+        ))}
+      </ul>
+      {persist && !readonly ? (
+        <p
+          className='alert alert-secondary'
+          role='alert'
+          data-testid='persistAlert'
+        >
+          <strong>NOTE:</strong> Feature flag values are persisting on page
+          refresh. This is not recommended for a production environment. Check
+          the <code>persist</code> setting when calling{' '}
+          <code>loadFeatureFlags</code> or <code>loadFeatureFlagsRedux</code>.
+        </p>
+      ) : null}
+      {!readonly ? (
+        <button
+          type='button'
+          className='btn btn-success'
+          data-testid='resetButton'
+          onClick={() => {
+            resetFlags();
+            onFeatureReset();
+          }}
+        >
+          Reset flags to default
+        </button>
+      ) : null}
     </>
   );
 };
 
-export default FeatureFlagsUI;
+export default CoreUI;
