+++ src/pages/ReadOnly.tsx
@@ -1,12 +1,12 @@
 import React, { ReactElement } from 'react';
-import FeatureFlagsContextUI from '../components/FeatureFlagsUI';
+import { FeatureFlagsUI } from '../components';
 
 //type ReadOnlyPageProps = {};
 
 const ReadOnlyPage = (): ReactElement => (
   <div className='container'>
     <h1>Local Storage Flags - Read Only</h1>
-    <FeatureFlagsContextUI readonly />
+    <FeatureFlagsUI readonly />
   </div>
 );
